requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
PyJWT
